Search for CIS studies.

Searches for CIS studies using the CIS search engine.

Usage

search_studies(
  search_terms = NULL,
  category_code = NULL,
  collection_code = NULL,
  since_date = NULL,
  until_date = NULL,
  study_code = NULL,
  since_study_code = NULL,
  until_study_code = NULL,
  scope_code = NULL
)

Arguments

search_terms

A string with the search terms.

category_code

A string (or a character vector if more than one) with the category code. The category code corresponds to the values of the codigo column in the data.frame output of the get_study_categories() function.

collection_code

A string (or a character vector if more than one) with the collection code. The collection code corresponds to the values of the codigo column in the data.frame output of the get_study_colections() function.

since_date

A string with the start date of the search in '%d-%m-%Y' format.

until_date

A string with the end date of the search in '%d-%m-%Y' format.

study_code

A string with the study code wanted.

since_study_code

A string with the first study code wanted.

until_study_code

A string with the last study code wanted.

scope_code

Scope level code of the survey sample ("001" for national level, "002" for autonomical comunity level, etc...)

Value

A data.frame.

Examples

if(Sys.info()["sysname"] == "Linux") httr::set_config(httr::config(ssl_verifypeer = 0L))
# Search by search terms
studies <- search_studies(search_terms = "postelectoral")
print(studies)
#> # A tibble: 179 × 4
#>       id codigo titulo                                                fecha     
#>    <int> <chr>  <chr>                                                 <date>    
#>  1 14752 3420   BARÓMETRO DE SEPTIEMBRE 2023. POSTELECTORAL ELECCION… 2023-09-01
#>  2 14723 3410   BARÓMETRO DE JUNIO 2023. POSTELECTORAL ELECCIONES MU… 2023-06-01
#>  3 14642 3370   POSTELECTORAL ELECCIONES AUTONÓMICAS 2022. COMUNIDAD… 2022-06-21
#>  4 14614 3352   POSTELECTORAL ELECCIONES AUTONÓMICAS 2022. COMUNIDAD… 2022-02-14
#>  5 14574 3328   POSTELECTORAL ELECCIONES AUTONÓMICAS 2021. COMUNIDAD… 2021-06-18
#>  6 14554 3314   POSTELECTORAL DE CATALUÑA. ELECCIONES AUTONÓMICAS 20… 2021-03-26
#>  7 14522 3293   POSTELECTORAL DEL PAÍS VASCO. ELECCIONES AUTONÓMICAS… 2020-09-08
#>  8 14523 3294   POSTELECTORAL DE GALICIA. ELECCIONES AUTONÓMICAS 2020 2020-09-08
#>  9 14479 3269   BARÓMETRO DE DICIEMBRE 2019. POSTELECTORAL ELECCIONE… 2019-12-01
#> 10 14455 3253   POSTELECTORAL ELECCIONES AUTONÓMICAS Y MUNICIPALES 2… 2019-06-17
#> # ℹ 169 more rows

# Narrow the search by dates
studies <- search_studies(search_terms = "postelectoral",
                          since_date = "01-01-2011",
                          until_date = "01-01-2020")
print(studies)
#> # A tibble: 53 × 4
#>       id codigo titulo                                                fecha     
#>    <int> <chr>  <chr>                                                 <date>    
#>  1 14479 3269   BARÓMETRO DE DICIEMBRE 2019. POSTELECTORAL ELECCIONE… 2019-12-01
#>  2 14455 3253   POSTELECTORAL ELECCIONES AUTONÓMICAS Y MUNICIPALES 2… 2019-06-17
#>  3 14453 3248   POSTELECTORAL ELECCIONES GENERALES 2019               2019-05-10
#>  4 14440 3236   POSTELECTORAL ELECCIONES AUTONÓMICAS 2018. COMUNIDAD… 2018-12-10
#>  5 14378 3202   POSTELECTORAL DE CATALUÑA. ELECCIONES AUTONÓMICAS 20… 2017-12-28
#>  6 14309 3154   POSTELECTORAL DEL PAÍS VASCO. ELECCIONES AUTONÓMICAS… 2016-09-28
#>  7 14310 3155   POSTELECTORAL DE GALICIA. ELECCIONES AUTONÓMICAS 2016 2016-09-28
#>  8 14291 3145   POSTELECTORAL ELECCIONES GENERALES 2016               2016-07-04
#>  9 14258 3126   POSTELECTORAL ELECCIONES GENERALES 2015. PANEL (2ª F… 2016-01-07
#> 10 14268 7715   PREELECTORAL Y POSTELECTORAL ELECCIONES GENERALES 20… 2015-10-27
#> # ℹ 43 more rows

# Use the scope_code parameter to search only national studies
studies <- search_studies(search_terms = "postelectoral",
                          since_date = "01-01-2011",
                          until_date = "01-01-2020",
                          scope_code = "001")
print(studies)
#> # A tibble: 9 × 4
#>      id codigo titulo                                                 fecha     
#>   <int> <chr>  <chr>                                                  <date>    
#> 1 14479 3269   BARÓMETRO DE DICIEMBRE 2019. POSTELECTORAL ELECCIONES… 2019-12-01
#> 2 14455 3253   POSTELECTORAL ELECCIONES AUTONÓMICAS Y MUNICIPALES 20… 2019-06-17
#> 3 14453 3248   POSTELECTORAL ELECCIONES GENERALES 2019                2019-05-10
#> 4 14291 3145   POSTELECTORAL ELECCIONES GENERALES 2016                2016-07-04
#> 5 14258 3126   POSTELECTORAL ELECCIONES GENERALES 2015. PANEL (2ª FA… 2016-01-07
#> 6 14268 7715   PREELECTORAL Y POSTELECTORAL ELECCIONES GENERALES 201… 2015-10-27
#> 7 14083 3028   POSTELECTORAL ELECCIONES AL PARLAMENTO EUROPEO 2014    2014-05-29
#> 8 12604 2920   POSTELECTORAL ELECCIONES GENERALES 2011. PANEL (2ª FA… 2011-11-24
#> 9 12644 7711   PREELECTORAL Y POSTELECTORAL ELECCIONES GENERALES 201… 2011-10-06